Fix #11
Numerai can now withdraw NMR sent to account numbers below 1,000,000.
This is safe because no one will be able to find the keys to these
addresses, and it will allow us to identify which data scientist sent
it to us, allow Numerai to stake those NMR on that data scientist’s
behalf.