Fix #9, Fix #5
New function stake() can be called by any address to make a stake on
itself.  Stakes made by Numerai on behalf of someone else are made with
stakeOnBehalf().  Timestamp is now recorded for each stake so that it’s
possible to enforce stake resolution time period.  Because the address
of staker is stored, it’s also possible to send ether back to the
staker in releaseStake() #6 in the future.